The web application
Here is the web application
We have 2 tabs
The first one show the list of the iBeacons in a table (loaded from a
json file). The "delete" button isn't complete because, sometimes, the
function add an index in the Json file.
The second tab is a form to add an iBeacon to the list. It's not
complete but the function is approximatly the same as "deleteBeacon"